@maybephilipp 2024-06-11
Hm, will look at it in the morning. Tests were padding on my local using node v20. Will check & fix