@steambap 2024-06-11
Your fix allows some crazy router definitions like:
```
/cmd/a_:tool
/cmd/:tool/:sub
/cmd/:tool/
```