@maybephilipp 2024-06-13
> Your fix allows some crazy router definitions like

@steambap Could you please explain why the routes are crazy? `a_:tool` is more exact than `:tool` and has a different prefix, so it's completely ok. The order of defining two last routes doesn't matter since ends up to be the same tree.

I've rolled out few fixes:
1. Fixed code so that failing test is passing.
2. Rollbacked wildcard duplicates check (I messed it up last time)
3. Improved wildcard tests to check equality of matched handles (to be sure that found route is the one that was supposed to be found)