@steambap 2024-06-13
What about routers like this:
```
/cmd/a_:tool
/cmd/ab_:tool
```

It is true that `a_:tool` is more exact than `:tool`. However, I prefer a solution that only allows either exact path or wildcard. There shouldn't be 2 wildcards.