@maybephilipp 2024-06-13
> However, I prefer a solution that only allows either exact path or wildcard. There shouldn't be 2 wildcards.

Okay, got it. What is the reasoning behind this? Why won't we allow multiple wildcards if they have a different prefix? It's the same performance for `search` since the prefix narrows down the route path. Could you please explain what are bad consequences of having it?