@steambap 2024-06-14
I think it's confusing. There are probably multiple ways to make route paths harder to read like:
```
/info/user/:public
/info/:user/public
```


Also, you mentioned performance. Backtrack search in this PR reduce performance by 1/10 ~ 1/3 in the router benchmark linked in readme, although the benchmark does not test anything that requires backtrack.