@maybephilipp 2024-06-14
> There are probably multiple ways to make route paths harder to read like

I agree that these routes are confusing, but I think it's 100% in the field of a tool user's responsibility, same as following correct REST route practices etc. It's all about code quality which must not be fixed by tools. Tools must provide functionality and I don't see a reason restricting it since exact/wildcard routes matching in similar paths is supported in most of well-known routers.

> Backtrack search in this PR reduce performance by 1/10 ~ 1/3

Oh, got it. Thanks for noticing. I will review performance and make changes then.