@maybephilipp 2024-06-14
@steambap I have done some optimizations, significantly improving performance. Now it runs with the same or +- same performance.

macOS, M1 Pro (arm), Node v20

## Before fix:

First run:

```
short static: 39,148,776 ops/sec
static with same radix: 16,651,172 ops/sec
dynamic route: 8,349,365 ops/sec
mixed static dynamic: 10,523,080 ops/sec
long static: 17,918,632 ops/sec
wildcard: 10,751,185 ops/sec
all together: 2,139,514 ops/sec
```

Second run:

```
short static: 39,988,004 ops/sec
static with same radix: 16,642,120 ops/sec
dynamic route: 8,342,339 ops/sec
mixed static dynamic: 10,567,501 ops/sec
long static: 17,936,764 ops/sec
wildcard: 10,537,075 ops/sec
all together: 2,267,038 ops/sec
```

## After fix

```
short static: 39,804,230 ops/sec
static with same radix: 16,203,987 ops/sec
dynamic route: 7,787,487 ops/sec
mixed static dynamic: 10,000,538 ops/sec
long static: 17,007,767 ops/sec
wildcard: 10,268,140 ops/sec
all together: 2,168,849 ops/sec
```

```
short static: 40,279,540 ops/sec
static with same radix: 16,566,337 ops/sec
dynamic route: 8,094,932 ops/sec
mixed static dynamic: 10,459,490 ops/sec
long static: 18,006,230 ops/sec
wildcard: 10,525,970 ops/sec
all together: 2,245,313 ops/sec
```