@steambap 2024-06-15
If this is a regex based router, I agree with you. We should provide user the tool to write `/:totally{-:insane}?{-:path}?`, and it's the user's responsibility to make sure they are correct.
However, for this module, I want to stay the same as the original Golang implementation. No multiple wildcards on the same route please.